Convert theme to checkoutHeader theme
These changes convert the used design of the last step including postal selection or birthdate form to the same theme that is used throughout the checkout process.
This allows for a better user experience.
Additionally pressing the text from radios on the invoice delivery method also swaps the checked radio.